+++ .github/workflows/workflow.yml
@@ -1,0 +1,31 @@
+name: CI  
+
+on:   
+  push:  
+    branches:  
+      - master  # Replace with your main branch name if different  
+  pull_request:  
+    branches:  
+      - master  # This runs on pull requests targeting the main branch  
+
+jobs:  
+  test:  
+    runs-on: ubuntu-latest  
+
+    steps:  
+      - name: Checkout repository  
+        uses: actions/checkout@v2  
+
+      - name: Set up Python  
+        uses: actions/setup-python@v3  
+        with:  
+          python-version: '3.12'  # You can specify the version of Python to use  
+
+      - name: Install dependencies  
+        run: |  
+          python -m pip install --upgrade pip  
+          pip install -r requirements.txt  
+
+      - name: Run tests  
+        run: |  
+          pytest  # Adjust this based on how you run your tests
