+++ src/driver_setup.py
@@ -1,0 +1,21 @@
+# driver_setup.py
+from selenium import webdriver
+from selenium.webdriver.chrome.service import Service
+from selenium.webdriver.chrome.options import Options
+
+def setup_driver():
+    options = Options()
+    options.add_argument('--disable-blink-features=AutomationControlled')
+    options.add_argument('--start-maximized')
+    options.add_experimental_option("prefs", {
+        "profile.default_content_setting_values.media_stream_mic": 1,
+        "profile.default_content_setting_values.media_stream_camera": 1,
+        "profile.default_content_setting_values.geolocation": 0,
+        "profile.default_content_setting_values.notifications": 1
+    })
+    
+    # You can also specify the path to your ChromeDriver if needed
+    # service = Service('/path/to/chromedriver')
+    # return webdriver.Chrome(service=service, options=options)
+    
+    return webdriver.Chrome(options=options)
